Revert "Feature/sign up with email"
