+++ lib/app.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'features/auth/Log in or sign with email/presentation/view/log_in_or_sign_up_with_email_view.dart';
+import 'features/auth/Log in or create an account screen/presentation/view/log_in_or_sign_up_with_email_view.dart';
 import 'features/auth/choose the registration method screen/presentation/view/choose_the_registration_method_screen.dart';
-import 'features/auth/sign_up/presentation/view/sign_up_view.dart';
 class TaskManagement extends StatelessWidget {
   const TaskManagement({super.key});
 
@@ -13,8 +12,7 @@ class TaskManagement extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       routes: {
         ChooseTheRegistrationMethodScreen.routeName: (context)=>const ChooseTheRegistrationMethodScreen(),
-        LogInOrSignUpWithEmailView.routeName: (context)=> const LogInOrSignUpWithEmailView(),
-        SignUpView.routeName:(context) => const SignUpView()
+        LogInOrSignUpWithEmailView.routeName: (context)=> const LogInOrSignUpWithEmailView()
       },
       initialRoute: ChooseTheRegistrationMethodScreen.routeName,
       localizationsDelegates: const [
