+++ lib/features/auth/Log in or create an account screen/presentation/view/log_in_or_sign_up_with_email_view.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:task_management_application/core/style/colors/application_color.dart';
 import 'package:task_management_application/core/utils/widget/app_bar/custom_leading_item.dart';
-import 'package:task_management_application/features/auth/Log%20in%20or%20sign%20with%20email/presentation/view/widget/custom_animated_button.dart';
+import 'package:task_management_application/features/auth/Log%20in%20or%20create%20an%20account%20screen/presentation/view/widget/custom_animated_button.dart';
 import '../../../choose the registration method screen/presentation/view/choose_the_registration_method_screen.dart';
-import '../../../sign_up/presentation/view/sign_up_view.dart';
 class LogInOrSignUpWithEmailView extends StatelessWidget {
   const LogInOrSignUpWithEmailView({super.key});
 
@@ -43,11 +42,11 @@ class LogInOrSignUpWithEmailView extends StatelessWidget {
             CustomAnimatedButton(
               pressEven: () {
                 //Navigator.pushNamed(context, "LoginView.routeName");
-              },textButton: "Log in",
+              },textButton: "Login",
             ),
             const SizedBox(height: 10),
             CustomAnimatedButton(pressEven: () {
-              Navigator.pushNamed(context, SignUpView.routeName);
+              //Navigator.pushNamed(context, "SignUpView.routeName");
             },textButton: "Sign Up",),
             SizedBox(height: MediaQuery.of(context).size.height * 0.04),
           ],
