+++ lib/features/auth/Log in or create an account screen/presentation/view/widget/custom_animated_button.dart
@@ -16,7 +16,7 @@ class CustomAnimatedButton extends StatelessWidget {
         },
         text: textButton,
         color: primaryColor,
-        buttonTextStyle: Theme.of(context).textTheme.titleLarge?.copyWith(color: Colors.white)
+        buttonTextStyle: Theme.of(context).textTheme.titleLarge
     );
   }
 }
