+++ lib/features/auth/choose the registration method screen/data/models/user_information_model.dart
@@ -3,20 +3,23 @@ class UserInformationModel{
   String? email;
   static const collectionName = "users";
   bool? isAccountSocial;
+  String? phoneNumber;
   String? id;
   String? fullName;
-  UserInformationModel({ this.fullName, this.email, this.id,this.pathImage,this.isAccountSocial});
+  UserInformationModel({ this.fullName, this.email, this.id,this.phoneNumber,this.pathImage,this.isAccountSocial});
   UserInformationModel.formFireStore(Map<String,dynamic>?data):this(
     pathImage: data?['pathImage'],
     email: data?['email'],
     id: data?['id'],
     isAccountSocial: data?['isAccountSocial'],
     fullName: data?['name'],
+    phoneNumber: data?["phone"],
   );
   Map<String,dynamic>toFireStore(){
     return {
       "name":fullName,
       "pathImage":pathImage,
+      "phone":phoneNumber,
       "email":email,
       "id":id,
       "isAccountSocial":isAccountSocial,
@@ -25,6 +28,6 @@ class UserInformationModel{
 
   @override
   String toString() {
-    return 'InformationUserModel{fullName: $fullName, email: $email, pathImage: $pathImage}';
+    return 'InformationUserModel{fullName: $fullName, email: $email, phoneNumber: $phoneNumber, pathImage: $pathImage}';
   }
 }
