+++ lib/features/auth/choose the registration method screen/presentation/view/choose_the_registration_method_screen.dart
@@ -31,5 +31,6 @@ class _ChooseTheRegistrationMethodScreenState extends BaseState<ChooseTheRegistr
   }
 
 
+  
 
 }
