+++ lib/features/auth/choose the registration method screen/presentation/view/widgets/screen_the_registration_method_body.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:provider/provider.dart';
-import '../../../../Log in or sign with email/presentation/view/log_in_or_sign_up_with_email_view.dart';
+import '../../../../Log in or create an account screen/presentation/view/log_in_or_sign_up_with_email_view.dart';
 import '../../view model/choose_the_registration_method_view_model.dart';
 import 'button_continue_account.dart';
 import 'custom_horizontal_line.dart';
