--- lib/features/auth/sign_up/presentation/view/widget/custom_leading_app_bar.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CustomLeadingAppBar extends StatelessWidget {
-  const CustomLeadingAppBar({super.key, this.onTap, required this.iconData});
-  final void Function()? onTap;
-  final IconData iconData;
-  @override
-  Widget build(BuildContext context) {
-    return IconButton(
-      onPressed: () {
-        onTap!();
-      }, icon: Icon(iconData) ,
-    );
-  }
-}
