--- lib/features/auth/sign_up/presentation/view/widget/display_message_If_Image_Not_Found.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class DisplayMessageIfImageNotFound extends StatelessWidget {
-  const DisplayMessageIfImageNotFound({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(vertical: 15.0),
-        child: Text("Please Enter Your Image",style: Theme.of(context).textTheme.bodyMedium?.copyWith(color: Colors.red),),
-      ),
-    );
-  }
-}
