--- lib/features/auth/sign_up/presentation/view_model/sign_up_navigator.dart
@@ -1,4 +1,0 @@
-import 'package:task_management_application/core/base/base_navigator.dart';
-
-abstract class SignUpNavigator extends BaseNavigator{
-}
