+++ pubspec.yaml
@@ -42,9 +42,6 @@ dependencies:
   intl: ^0.19.0
   cached_network_image: ^3.3.1
   awesome_dialog: ^3.2.1
-  intl_phone_field:
-    git:
-      url: https://github.com/HalimWagih66/intl_phone_field.git
   provider: ^6.1.2
   fluttertoast: ^8.2.6
   google_sign_in: ^6.2.1
